Makes the code more testable.
- add the interface `Sieve` matching the original `Filter`'s interface

- move line filtering functionality into their own class, `LineFilter`, which implements the interface `LineSieve`
- move the tests for line filtering to `LineFilterTest` (tests are also reformatted a bit)
- the public interface is 100% preserved, existing methods delegate to `$this->lineFilter->*` functions, which are essentially the same.

- abstraction for include_once() in initializeData()
