+++ src/CodeCoverage.php
@@ -32,7 +32,7 @@ final class CodeCoverage
     private $driver;
 
     /**
-     * @var Filter
+     * @var Sieve
      */
     private $filter;
 
@@ -40,11 +40,6 @@ final class CodeCoverage
      * @var Wizard
      */
     private $wizard;
-
-    /**
-     * @var bool
-     */
-    private $cacheTokens = false;
 
     /**
      * @var bool
@@ -77,11 +72,6 @@ final class CodeCoverage
     private $processUncoveredFilesFromWhitelist = false;
 
     /**
-     * @var bool
-     */
-    private $ignoreDeprecatedCode = false;
-
-    /**
      * @var PhptTestCase|string|TestCase
      */
     private $currentId;
@@ -92,16 +82,6 @@ final class CodeCoverage
      * @var array
      */
     private $data = [];
-
-    /**
-     * @var array
-     */
-    private $ignoredLines = [];
-
-    /**
-     * @var bool
-     */
-    private $disableIgnoredLines = false;
 
     /**
      * Test data.
@@ -135,9 +115,24 @@ final class CodeCoverage
     private $report;
 
     /**
+     * @var \Closure
+     */
+    private $evaluateWhitelistedFile;
+
+    /**
+     * @var LineSieve
+     */
+    private $lineFilter;
+
+    /**
      * @throws RuntimeException
      */
-    public function __construct(Driver $driver = null, Filter $filter = null)
+    public function __construct(
+        Driver $driver = null,
+        Sieve $filter = null,
+        \Closure $evaluateWhitelistedFile = null,
+        LineSieve $lineFilter = null
+    )
     {
         if ($filter === null) {
             $filter = new Filter;
@@ -147,8 +142,20 @@ final class CodeCoverage
             $driver = $this->selectDriver($filter);
         }
 
-        $this->driver = $driver;
-        $this->filter = $filter;
+        if ($evaluateWhitelistedFile === null) {
+            $evaluateWhitelistedFile = function(string $file){
+                include_once $file;
+            };
+        }
+
+        if ($lineFilter === null) {
+            $lineFilter = new LineFilter();
+        }
+
+        $this->driver                  = $driver;
+        $this->filter                  = $filter;
+        $this->evaluateWhitelistedFile = $evaluateWhitelistedFile;
+        $this->lineFilter              = $lineFilter;
 
         $this->wizard = new Wizard;
     }
@@ -182,7 +189,7 @@ final class CodeCoverage
     /**
      * Returns the filter object used.
      */
-    public function filter(): Filter
+    public function filter(): Sieve
     {
         return $this->filter;
     }
@@ -406,12 +413,12 @@ final class CodeCoverage
 
     public function setCacheTokens(bool $flag): void
     {
-        $this->cacheTokens = $flag;
+        $this->lineFilter->setCacheTokens($flag);
     }
 
     public function getCacheTokens(): bool
     {
-        return $this->cacheTokens;
+        return $this->lineFilter->getCacheTokens();
     }
 
     public function setCheckForUnintentionallyCoveredCode(bool $flag): void
@@ -446,12 +453,12 @@ final class CodeCoverage
 
     public function setDisableIgnoredLines(bool $flag): void
     {
-        $this->disableIgnoredLines = $flag;
+        $this->lineFilter->setDisableIgnoredLines($flag);
     }
 
     public function setIgnoreDeprecatedCode(bool $flag): void
     {
-        $this->ignoreDeprecatedCode = $flag;
+        $this->lineFilter->setIgnoreDeprecatedCode($flag);
     }
 
     public function setUnintentionallyCoveredSubclassesWhitelist(array $whitelist): void
@@ -555,7 +562,7 @@ final class CodeCoverage
                 continue;
             }
 
-            foreach ($this->getLinesToBeIgnored($filename) as $line) {
+            foreach ($this->lineFilter->getLinesToBeIgnored($filename) as $line) {
                 unset($data[$filename][$line]);
             }
         }
@@ -606,190 +613,6 @@ final class CodeCoverage
         }
 
         $this->append($data, 'UNCOVERED_FILES_FROM_WHITELIST');
-    }
-
-    private function getLinesToBeIgnored(string $fileName): array
-    {
-        if (isset($this->ignoredLines[$fileName])) {
-            return $this->ignoredLines[$fileName];
-        }
-
-        try {
-            return $this->getLinesToBeIgnoredInner($fileName);
-        } catch (\OutOfBoundsException $e) {
-            // This can happen with PHP_Token_Stream if the file is syntactically invalid,
-            // and probably affects a file that wasn't executed.
-            return [];
-        }
-    }
-
-    private function getLinesToBeIgnoredInner(string $fileName): array
-    {
-        $this->ignoredLines[$fileName] = [];
-
-        $lines = \file($fileName);
-
-        foreach ($lines as $index => $line) {
-            if (!\trim($line)) {
-                $this->ignoredLines[$fileName][] = $index + 1;
-            }
-        }
-
-        if ($this->cacheTokens) {
-            $tokens = \PHP_Token_Stream_CachingFactory::get($fileName);
-        } else {
-            $tokens = new \PHP_Token_Stream($fileName);
-        }
-
-        foreach ($tokens->getInterfaces() as $interface) {
-            $interfaceStartLine = $interface['startLine'];
-            $interfaceEndLine   = $interface['endLine'];
-
-            foreach (\range($interfaceStartLine, $interfaceEndLine) as $line) {
-                $this->ignoredLines[$fileName][] = $line;
-            }
-        }
-
-        foreach (\array_merge($tokens->getClasses(), $tokens->getTraits()) as $classOrTrait) {
-            $classOrTraitStartLine = $classOrTrait['startLine'];
-            $classOrTraitEndLine   = $classOrTrait['endLine'];
-
-            if (empty($classOrTrait['methods'])) {
-                foreach (\range($classOrTraitStartLine, $classOrTraitEndLine) as $line) {
-                    $this->ignoredLines[$fileName][] = $line;
-                }
-
-                continue;
-            }
-
-            $firstMethod          = \array_shift($classOrTrait['methods']);
-            $firstMethodStartLine = $firstMethod['startLine'];
-            $firstMethodEndLine   = $firstMethod['endLine'];
-            $lastMethodEndLine    = $firstMethodEndLine;
-
-            do {
-                $lastMethod = \array_pop($classOrTrait['methods']);
-            } while ($lastMethod !== null && 0 === \strpos($lastMethod['signature'], 'anonymousFunction'));
-
-            if ($lastMethod !== null) {
-                $lastMethodEndLine = $lastMethod['endLine'];
-            }
-
-            foreach (\range($classOrTraitStartLine, $firstMethodStartLine) as $line) {
-                $this->ignoredLines[$fileName][] = $line;
-            }
-
-            foreach (\range($lastMethodEndLine + 1, $classOrTraitEndLine) as $line) {
-                $this->ignoredLines[$fileName][] = $line;
-            }
-        }
-
-        if ($this->disableIgnoredLines) {
-            $this->ignoredLines[$fileName] = \array_unique($this->ignoredLines[$fileName]);
-            \sort($this->ignoredLines[$fileName]);
-
-            return $this->ignoredLines[$fileName];
-        }
-
-        $ignore = false;
-        $stop   = false;
-
-        foreach ($tokens->tokens() as $token) {
-            switch (\get_class($token)) {
-                case \PHP_Token_COMMENT::class:
-                case \PHP_Token_DOC_COMMENT::class:
-                    $_token = \trim($token);
-                    $_line  = \trim($lines[$token->getLine() - 1]);
-
-                    if ($_token === '// @codeCoverageIgnore' ||
-                        $_token === '//@codeCoverageIgnore') {
-                        $ignore = true;
-                        $stop   = true;
-                    } elseif ($_token === '// @codeCoverageIgnoreStart' ||
-                        $_token === '//@codeCoverageIgnoreStart') {
-                        $ignore = true;
-                    } elseif ($_token === '// @codeCoverageIgnoreEnd' ||
-                        $_token === '//@codeCoverageIgnoreEnd') {
-                        $stop = true;
-                    }
-
-                    if (!$ignore) {
-                        $start = $token->getLine();
-                        $end   = $start + \substr_count($token, "\n");
-
-                        // Do not ignore the first line when there is a token
-                        // before the comment
-                        if (0 !== \strpos($_token, $_line)) {
-                            $start++;
-                        }
-
-                        for ($i = $start; $i < $end; $i++) {
-                            $this->ignoredLines[$fileName][] = $i;
-                        }
-
-                        // A DOC_COMMENT token or a COMMENT token starting with "/*"
-                        // does not contain the final \n character in its text
-                        if (isset($lines[$i - 1]) && 0 === \strpos($_token, '/*') && '*/' === \substr(\trim($lines[$i - 1]), -2)) {
-                            $this->ignoredLines[$fileName][] = $i;
-                        }
-                    }
-
-                    break;
-
-                case \PHP_Token_INTERFACE::class:
-                case \PHP_Token_TRAIT::class:
-                case \PHP_Token_CLASS::class:
-                case \PHP_Token_FUNCTION::class:
-                    /* @var \PHP_Token_Interface $token */
-
-                    $docblock = $token->getDocblock();
-
-                    $this->ignoredLines[$fileName][] = $token->getLine();
-
-                    if (\strpos($docblock, '@codeCoverageIgnore') || ($this->ignoreDeprecatedCode && \strpos($docblock, '@deprecated'))) {
-                        $endLine = $token->getEndLine();
-
-                        for ($i = $token->getLine(); $i <= $endLine; $i++) {
-                            $this->ignoredLines[$fileName][] = $i;
-                        }
-                    }
-
-                    break;
-
-                /* @noinspection PhpMissingBreakStatementInspection */
-                case \PHP_Token_NAMESPACE::class:
-                    $this->ignoredLines[$fileName][] = $token->getEndLine();
-
-                // Intentional fallthrough
-                case \PHP_Token_DECLARE::class:
-                case \PHP_Token_OPEN_TAG::class:
-                case \PHP_Token_CLOSE_TAG::class:
-                case \PHP_Token_USE::class:
-                    $this->ignoredLines[$fileName][] = $token->getLine();
-
-                    break;
-            }
-
-            if ($ignore) {
-                $this->ignoredLines[$fileName][] = $token->getLine();
-
-                if ($stop) {
-                    $ignore = false;
-                    $stop   = false;
-                }
-            }
-        }
-
-        $this->ignoredLines[$fileName][] = \count($lines) + 1;
-
-        $this->ignoredLines[$fileName] = \array_unique(
-            $this->ignoredLines[$fileName]
-        );
-
-        $this->ignoredLines[$fileName] = \array_unique($this->ignoredLines[$fileName]);
-        \sort($this->ignoredLines[$fileName]);
-
-        return $this->ignoredLines[$fileName];
     }
 
     /**
@@ -894,7 +717,7 @@ final class CodeCoverage
     /**
      * @throws RuntimeException
      */
-    private function selectDriver(Filter $filter): Driver
+    private function selectDriver(Sieve $filter): Driver
     {
         $runtime = new Runtime;
 
@@ -955,16 +778,14 @@ final class CodeCoverage
             $this->shouldCheckForDeadAndUnused = false;
 
             $this->driver->start();
-
             foreach ($this->filter->getWhitelist() as $file) {
                 if ($this->filter->isFile($file)) {
-                    include_once $file;
+                    ($this->evaluateWhitelistedFile)($file);
                 }
             }
-
-            $data     = [];
             $coverage = $this->driver->stop();
 
+            $data     = [];
             foreach ($coverage as $file => $lines) {
                 if ($this->filter->isFiltered($file)) {
                     continue;
